+++ benchmarks/Compression/Program.cs
@@ -31,8 +31,6 @@ static void OutputCompressionInfo(MessageSize size, MessageType type)
         Console.WriteLine($"\tCompressed with K4os.Compression.LZ4: {compressor.Compress(data).Length}");
     using (var compressor = Factories.IronSnappyCompressorFactory.Create())
         Console.WriteLine($"\tCompressed with IronSnappy: {compressor.Compress(data).Length}");
-    using (var compressor = Factories.DotNetZipCompressorFactory.Create())
-        Console.WriteLine($"\tCompressed with DotNetZip: {compressor.Compress(data).Length}");
     using (var compressor = Factories.BuiltinZlibCompressionCompressorFactory.Create())
         Console.WriteLine($"\tCompressed with System.IO.Compression: {compressor.Compress(data).Length}");
     using (var compressor = Factories.ZstdNetCompressorFactory.Create())